requirements.txt
httpx~=0.25.0
setuptools~=49.2.1
pytest~=7.4.2
pydantic~=1.10.12
python-dotenv~=1.0.0